X-Ray_Workbench_AI/Chest_xray/requirements.txt
Flask
flask-cors
tensorflow==2.13.0
Pillow
numpy
keras==2.13.1